requirements.txt
appdirs==1.4.3
backcall==0.1.0
colorama==0.4.3
decorator==4.4.1
distlib==0.3.0
filelock==3.0.12
gitdb2==2.0.6
GitPython==3.0.7
ipdb==0.12.3
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.1
pickleshare==0.7.5
prompt-toolkit==3.0.3
Pygments==2.5.2
six==1.14.0
smmap2==2.0.5
traitlets==4.3.3
virtualenv==20.0.4
wcwidth==0.1.8
